Removing manager from init
