+++ tests/test_manager.py
